// ts/src/generated/schema.ts
export interface components {
    schemas: {
// ...
        asset: {
            /** @enum {string} */
            archiveExtractionStatus?: "pending" | "in_progress" | "done" | "failed";
            contentType?: string;
            /** Format: date-time */
            createdAt: string;
            file?: components["schemas"]["file"];
            id: string;
            name?: string;
            /** @enum {string} */
            previewType?: "image" | "image_svg" | "geo" | "geo_3d_Tiles" | "geo_mvt" | "model_3d" | "csv" | "unknown";
            projectId: string;
            public: boolean;
            totalSize?: number;
            /** Format: date-time */
            updatedAt: string;
            url: string;
        };
// ...
        field: {
            group?: string;
            id?: string;
            key?: string;
            type?: components["schemas"]["valueType"];
            value?: unknown;
        };
// ...
        item: {
            /** Format: date-time */
            createdAt?: string;
            fields?: components["schemas"]["field"][];
            id?: string;
            isMetadata?: boolean;
            metadataFields?: components["schemas"]["field"][];
            metadataItemId?: string;
            modelId?: string;
            originalItemId?: string;
            referencedItems?: components["schemas"]["item"][];
            /** Format: date-time */
            updatedAt?: string;
        };
        model: {
            /** Format: date-time */
            createdAt?: string;
            description?: string;
            id?: string;
            key?: string;
            /** Format: date-time */
            lastModified?: string;
            metadataSchema?: components["schemas"]["schema"];
            metadataSchemaId?: string;
            name?: string;
            projectId?: string;
            schema?: components["schemas"]["schema"];
            schemaId?: string;
            /** Format: date-time */
            updatedAt?: string;
        };
// ...
    };
// ...
}